+++ release_notes.md
@@ -8,7 +8,10 @@ PREESM Changelog
 
 ### Changes
 * Fix Papi event add/remove;
+* Deprecated PapifyEngine task, replaced with the new parameter in the codegen task;
 * Update sonar plugin to 3.6.0.1398;
+* Fix Scenario UI layouts;
+* Refactoring;
 
 ### Bug fix
 * Fix #154: fix scenario UI layouts;